requirements.txt
django==3.0.5
djangorestframework==3.11.0
django-rest-knox==4.1.0
bcrypt
requests
channels==2.4.0
channels_redis
django_encrypted_filefield
python-magic-bin
